Single header breakdown and more view tests
