+++ include/moon/types.h
@@ -1,0 +1,18 @@
+#pragma once
+
+namespace moon {
+template <typename T>
+using LuaMap = std::unordered_map<std::string, T>;
+
+enum class LuaType {
+    Null = LUA_TNIL,
+    Boolean = LUA_TBOOLEAN,
+    LightUserData = LUA_TLIGHTUSERDATA,
+    Number = LUA_TNUMBER,
+    String = LUA_TSTRING,
+    Table = LUA_TTABLE,
+    Function = LUA_TFUNCTION,
+    UserData = LUA_TUSERDATA,
+    Thread = LUA_TTHREAD
+};
+}  // namespace moon
