+++ include/moon/usertype.h
@@ -1,0 +1,339 @@
+#pragma once
+
+namespace moon {
+/**
+ * @brief Converts C++ class to Lua metatable.
+ * LunaFive modded - http://lua-users.org/wiki/LunaFive
+ * This modded version uses a different method to bind class methods.
+ *
+ * @tparam T Class to expose to Lua.
+ */
+template <class T>
+class LuaClass {
+public:
+    struct PropertyType {
+        const char* name;
+
+        int (T::*getter)(lua_State*);
+
+        int (T::*setter)(lua_State*);
+    };
+
+    struct FunctionType {
+        const char* name;
+
+        int (T::*func)(lua_State*);
+    };
+
+    /**
+     * @brief Retrieves a wrapped class from the arguments passed to the func, specified by narg (position).
+     * This func will raise an exception if the argument is not of the correct type.
+     *
+     * @param L Lua State
+     * @param narg Position to check
+     * @return T*
+     */
+    static T* Check(lua_State* L, int narg) {
+        T** obj = static_cast<T**>(luaL_checkudata(L, narg, T::Binding.GetName()));
+        if (!obj) {
+            return nullptr;
+        }             // LightCheck returns nullptr if not found.
+        return *obj;  // pointer to T object
+    }
+
+    /**
+     * @brief Retrieves a wrapped class from the arguments passed to the func, specified by narg (position).
+     * This func will return nullptr if the argument is not of the correct type.  Useful for supporting
+     * multiple types of arguments passed to the func
+     *
+     * @param L Lua State
+     * @param narg Position to check
+     * @return T*
+     */
+    static T* LightCheck(lua_State* L, int narg) {
+        T** obj = static_cast<T**>(luaL_testudata(L, narg, T::Binding.GetName()));
+        if (!obj) {
+            return nullptr;
+        }             // LightCheck returns nullptr if not found.
+        return *obj;  // pointer to T object
+    }
+
+    /**
+     * @brief Registers your class with Lua.  Leave nameSpace "" if you want to load it into the global space.
+     *
+     * @param L Lua State
+     * @param nameSpace Namespace to load into
+     */
+    static void Register(lua_State* L, const char* nameSpace = nullptr) {
+        if (nameSpace != nullptr && strlen(nameSpace) != 0u) {
+            lua_getglobal(L, nameSpace);
+            // Create namespace if not present
+            if (lua_isnil(L, -1)) {
+                lua_pop(L, 1);
+                lua_newtable(L);
+                lua_pushvalue(L, -1);  // Duplicate table pointer since setglobal pops the value
+                lua_setglobal(L, nameSpace);
+            }
+            lua_pushcfunction(L, &LuaClass<T>::constructor);
+            lua_setfield(L, -2, T::Binding.GetName());
+            lua_pop(L, 1);
+        } else {
+            lua_pushcfunction(L, &LuaClass<T>::constructor);
+            lua_setglobal(L, T::Binding.GetName());
+        }
+
+        luaL_newmetatable(L, T::Binding.GetName());
+        int metatable = lua_gettop(L);
+
+        lua_pushstring(L, "__tostring");
+        lua_pushcfunction(L, &LuaClass<T>::to_string);
+        lua_settable(L, metatable);
+        // To be able to compare two Luna objects (not natively possible with full userdata)
+        lua_pushstring(L, "__eq");
+        lua_pushcfunction(L, &LuaClass<T>::equals);
+        lua_settable(L, metatable);
+
+        lua_pushstring(L, "__index");
+        lua_pushcfunction(L, &LuaClass<T>::property_getter);
+        lua_settable(L, metatable);
+
+        lua_pushstring(L, "__newindex");
+        lua_pushcfunction(L, &LuaClass<T>::property_setter);
+        lua_settable(L, metatable);
+
+        if (T::Binding.GetGC()) {
+            lua_pushstring(L, "__gc");
+            lua_pushcfunction(L, &LuaClass<T>::gc_obj);
+            lua_settable(L, metatable);
+        }
+
+        unsigned i = 0;
+        for (const auto& property : T::Binding.GetProperties()) {  // Register some properties in it
+            lua_pushstring(L, property.name);                      // Having some string associated with them
+            lua_pushnumber(L, i++);                                // And an index of which property it is
+            lua_settable(L, metatable);
+        }
+
+        i = 0;
+        for (const auto& method : T::Binding.GetMethods()) {
+            lua_pushstring(L, method.name);       // Register some functions in it
+            lua_pushnumber(L, i++ | (1u << 8u));  // Add a number indexing which func it is
+            lua_settable(L, metatable);
+        }
+        lua_pop(L, 1);
+    }
+
+    /**
+     * @brief Loads an instance of the class into the Lua stack, and provides you a pointer so you can modify it.
+     *
+     * @param L Lua State
+     * @param instance Instance to push
+     */
+    static void Push(lua_State* L, T* instance) {
+        T** a = static_cast<T**>(lua_newuserdata(L, sizeof(T*)));  // Create userdata
+        *a = instance;
+        luaL_getmetatable(L, T::Binding.GetName());
+        lua_setmetatable(L, -2);
+    }
+
+private:
+    /**
+     * @brief constructor (internal)
+     *
+     * @param L Lua State
+     * @return int
+     */
+    static int constructor(lua_State* L) {
+        T* ap = new T(L);
+        T** a = static_cast<T**>(lua_newuserdata(L, sizeof(T*)));  // Push value = userdata
+        *a = ap;
+
+        luaL_getmetatable(L, T::Binding.GetName());  // Fetch global metatable T::classname
+        lua_setmetatable(L, -2);
+        return 1;
+    }
+
+    /**
+     * @brief property_getter (internal)
+     *
+     * @param L Lua State
+     * @return int
+     */
+    static int property_getter(lua_State* L) {
+        lua_getmetatable(L, 1);  // Look up the index of a name
+        lua_pushvalue(L, 2);     // Push the name
+        lua_rawget(L, -2);       // Get the index
+
+        if (lua_isnumber(L, -1) != 0) {  // Check if we got a valid index
+            auto _index = static_cast<unsigned>(lua_tonumber(L, -1));
+
+            T** obj = static_cast<T**>(lua_touserdata(L, 1));
+
+            lua_pushvalue(L, 3);
+
+            // A func
+            if ((_index & (1u << 8u)) != 0u) {
+                lua_pushnumber(L, _index ^ (1u << 8u));  // Push the right func index
+                lua_pushlightuserdata(L, obj);
+                lua_pushcclosure(L, &LuaClass<T>::function_dispatch, 2);
+                return 1;  // Return a func
+            }
+
+            lua_pop(L, 2);     // Pop metatable and _index
+            lua_remove(L, 1);  // Remove userdata
+            lua_remove(L, 1);  // Remove [key]
+
+            return ((*obj)->*(T::Binding.GetProperties()[_index].getter))(L);
+        }
+
+        return 1;
+    }
+
+    /**
+     * @brief property_setter (internal)
+     *
+     * @param L Lua State
+     * @return int
+     */
+    static int property_setter(lua_State* L) {
+        lua_getmetatable(L, 1);  // Look up the index from name
+        lua_pushvalue(L, 2);     //
+        lua_rawget(L, -2);       //
+
+        if (lua_isnumber(L, -1) != 0) {  // Check if we got a valid index
+            auto _index = static_cast<unsigned>(lua_tonumber(L, -1));
+
+            T** obj = static_cast<T**>(lua_touserdata(L, 1));
+
+            if (!obj || !*obj) {
+                luaL_error(L, "Internal error, no object given!");
+                return 0;
+            }
+
+            if ((_index >> 8u) != 0u) {  // Try to set a func
+                char c[128];
+#if !defined(__MINGW32__) && (defined(_WIN32) || defined(__WIN32__) || defined(WIN32))
+                sprintf_s(c, "Moon: Trying to set the method [%s] of class [%s]", (*obj)->T::Binding.GetMethods()[_index ^ (1u << 8u)].name,
+                          T::Binding.GetName());
+#else
+                sprintf(c, "Moon: Trying to set the method [%s] of class [%s]", (*obj)->T::Binding.GetMethods()[_index ^ (1u << 8u)].name,
+                        T::Binding.GetName());
+#endif
+                luaL_error(L, c);
+                return 0;
+            }
+
+            lua_pop(L, 2);     // Pop metatable and _index
+            lua_remove(L, 1);  // Remove userdata
+            lua_remove(L, 1);  // Remove [key]
+
+            return ((*obj)->*(T::Binding.GetProperties()[_index].setter))(L);
+        }
+
+        return 0;
+    }
+
+    /**
+     * @brief function_dispatch (internal)
+     *
+     * @param L Lua State
+     * @return int
+     */
+    static int function_dispatch(lua_State* L) {
+        int i = (int)lua_tonumber(L, lua_upvalueindex(1));
+        T** obj = static_cast<T**>(lua_touserdata(L, lua_upvalueindex(2)));
+
+        return ((*obj)->*(T::Binding.GetMethods()[i].func))(L);
+    }
+
+    /**
+     * @brief gc_obj (internal)
+     *
+     * @param L Lua State
+     * @return int
+     */
+    static int gc_obj(lua_State* L) {
+        T** obj = static_cast<T**>(lua_touserdata(L, -1));
+
+        if (obj && *obj) {
+            delete (*obj);
+        }
+
+        return 0;
+    }
+
+    static int to_string(lua_State* L) {
+        T** obj = static_cast<T**>(lua_touserdata(L, -1));
+
+        if (obj) {
+            lua_pushfstring(L, "%s (%p)", T::Binding.GetName(), (void*)*obj);
+        } else {
+            lua_pushstring(L, "Empty object");
+        }
+
+        return 1;
+    }
+
+    /**
+     * @brief Method which compares two Luna objects.
+     * The full userdatas (as opposed to light userdata) can't be natively compared one to other, we have to had this to
+     * do it.
+     *
+     * @param L Lua State
+     * @return int
+     */
+    static int equals(lua_State* L) {
+        T** obj1 = static_cast<T**>(lua_touserdata(L, -1));
+        T** obj2 = static_cast<T**>(lua_touserdata(L, 1));
+
+        lua_pushboolean(L, *obj1 == *obj2);
+
+        return 1;
+    }
+};
+
+template <class BindableClass>
+class Binding {
+public:
+    using LuaFunction = typename LuaClass<BindableClass>::FunctionType;
+    using LuaProperty = typename LuaClass<BindableClass>::PropertyType;
+
+    explicit Binding(const char* name) : m_name(name) {}
+
+    ~Binding() = default;
+
+    [[nodiscard]] inline const char* GetName() const { return m_name; }
+
+    [[nodiscard]] inline const auto& GetMethods() const { return m_methods; }
+
+    [[nodiscard]] inline const auto& GetProperties() const { return m_properties; }
+
+    [[nodiscard]] inline bool GetGC() const { return m_gc; }
+
+    Binding& RemoveGC() {
+        m_gc = false;
+        return *this;
+    }
+
+    Binding& AddMethod(LuaFunction func) {
+        m_methods.push_back(func);
+        return *this;
+    }
+
+    Binding& AddProperty(LuaProperty prop) {
+        m_properties.push_back(prop);
+        return *this;
+    }
+
+private:
+    const char* m_name;
+    std::vector<LuaFunction> m_methods;
+    std::vector<LuaProperty> m_properties;
+    bool m_gc{true};
+};
+
+namespace meta {
+template <typename T, typename Ret = T>
+using is_binding_t = std::enable_if_t<std::is_same_v<decltype(T::Binding), Binding<T>>, Ret>;
+}
+}  // namespace moon
